+++ src/main/java/controller/ControllerHandler.java
@@ -2,9 +2,9 @@ package controller;
 
 import exception.ControllerNotFoundException;
 import http.request.HttpRequest;
-import http.HttpResponse;
-import http.request.RequestLine;
 import http.request.HttpUri;
+import http.request.RequestLine;
+import http.response.HttpResponse;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
